reset_and_reload.py
"""
ONE-SHOT COMPLETE RESET AND RELOAD
Drops all database tables, clears FAISS index, and reloads knowledge base fresh
"""

import os
import sys
import shutil
import logging
from pathlib import Path

# Setup logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

def reset_and_reload():
    """Complete reset workflow"""
    
    print("\n" + "="*80)
    print("COMPLETE RESET AND RELOAD - ONE SHOT FIX")
    print("="*80)
    
    # Step 1: Drop all database tables
    print("\n[STEP 1/4] Dropping all database tables...")
    try:
        from src.database.database import Base
        from sqlalchemy import create_engine, inspect, text
        from config.settings import get_settings
        
        settings = get_settings()
        engine = create_engine(settings.DATABASE_URL)
        
        # Drop all tables
        with engine.begin() as conn:
            inspector = inspect(conn)
            for table_name in reversed(inspector.get_table_names()):
                logger.info(f"  Dropping table: {table_name}")
                conn.execute(text(f"DROP TABLE IF EXISTS {table_name}"))
        
        # Recreate schema fresh
        logger.info("  Recreating schema fresh...")
        Base.metadata.create_all(engine)
        logger.info("✓ Database reset complete")
    except Exception as e:
        logger.error(f"✗ Database reset failed: {e}")
        return False
    
    # Step 2: Clear FAISS index
    print("\n[STEP 2/4] Clearing FAISS index...")
    try:
        faiss_dir = Path("data/vector_index")
        faiss_files = [
            faiss_dir / "faiss.index",
            faiss_dir / "chunks.index",
            faiss_dir / "faiss_metadata.pkl",
            faiss_dir / "chunks_metadata.pkl"
        ]
        
        for file_path in faiss_files:
            if file_path.exists():
                logger.info(f"  Deleting: {file_path.name}")
                os.remove(file_path)
        
        logger.info("✓ FAISS index cleared")
    except Exception as e:
        logger.error(f"✗ FAISS cleanup failed: {e}")
        return False
    
    # Step 3: Run knowledge base loader
    print("\n[STEP 3/4] Loading knowledge base fresh...")
    try:
        # Import after database reset
        from load_knowledge_base import main as load_kb
        
        logger.info(f"  Starting knowledge base load...")
        success = load_kb()
        
        if not success:
            logger.warning("✓ Knowledge base load completed with warnings")
        else:
            logger.info("✓ Knowledge base loaded successfully")
    except Exception as e:
        logger.error(f"✗ Knowledge base load failed: {e}")
        import traceback
        traceback.print_exc()
        return False
    
    # Step 4: Verify FAISS index
    print("\n[STEP 4/4] Verifying FAISS index...")
    try:
        from src.vector_db.index_manager import FAISSManager
        import faiss
        
        manager = FAISSManager()
        index_path = Path("data/vector_index/chunks.index")
        
        if index_path.exists():
            index = faiss.read_index(str(index_path))
            ntotal = index.ntotal
            logger.info(f"✓ FAISS index verified: {ntotal} vectors")
        else:
            logger.warning("⚠ FAISS index not found - will be created on first query")
    except Exception as e:
        logger.warning(f"⚠ Could not verify FAISS index: {e}")
    
    print("\n" + "="*80)
    print("✓ COMPLETE RESET AND RELOAD FINISHED")
    print("="*80)
    print("\nYour knowledge base is now:")
    print("  • Database: Fresh (all duplicates removed)")
    print("  • FAISS Index: Cleared and rebuilt")
    print("  • Status: Ready for production use")
    print("\n" + "="*80 + "\n")
    
    return True

if __name__ == "__main__":
    try:
        success = reset_and_reload()
        sys.exit(0 if success else 1)
    except KeyboardInterrupt:
        logger.warning("\nReset cancelled by user")
        sys.exit(1)
    except Exception as e:
        logger.error(f"Unexpected error: {e}")
        import traceback
        traceback.print_exc()
        sys.exit(1)
